requirements.txt
# Core AI and LLM dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# PDF processing
pypdf2>=3.0.1
pymupdf>=1.23.0
pdfplumber>=0.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Vector search
faiss-cpu>=1.7.4

# Arxiv integration
arxiv>=1.4.0
requests>=2.31.0

# Web interface (optional)
streamlit>=1.28.0

# Logging and utilities
structlog>=23.1.0
rich>=13.6.0

# Testing
pytest>=7.4.0
